Moved 'taskOperation' mock data into new mock file
